refactor: splitting classes by files
